pyproject.toml
[project]
name = "carecompanion"
version = "1.0.0"
description = "Voice-first senior care AI companion"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]",
    "uvicorn[standard]",
    "httpx",
    "pydantic-settings",
    "apscheduler>=3.10,<4",
]

[project.optional-dependencies]
test = ["pytest", "pytest-asyncio"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
